+++ test/client/functions/batch/hisRead.js
@@ -159,7 +159,7 @@ describe("client.batch.hisRead", () => {
             expect(result.errors.length).to.be.equal(1);
             const { error, args } = result.errors[0];
             expect(error).to.equal("Test error");
-            expect(args).to.eql([["id1"], "start", "end", 100]);
+            expect(args).to.eql(["hisRead", ["id1"], "start", "end", 100]);
         });
     });